Added test case for user-key generation.
Do not run the test if you do not want to test key generation, otherwise the test will fail specifying that the keys already exist. It may be advisable to change the test case to fail if keys do not exist at the end of the function. Or maybe this should be considered elsewhere.