KeyExists() function added, GenUserKeys() mod, userkey_test.go mod
KeyExists() is used in GenUserKeys() to ensure that keys are not overwritten, and that there is not a private key without a public key and vice versa. userkey_test.go will fail only if there is a missing single key, or if there is an error in generating a key. It will succeed if a key-pair is successfully generated. It will succeed if a key-pair exists, but will not overwrite the old pair; Instead it will notify that the key-pair already exists.